add some tests for registration cards
